+++ nym-vpn-core/crates/nym-ipc/src/authentication/macos.rs
@@ -2,13 +2,21 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 use tokio::net::UnixStream;
-use tokio_util::sync::CancellationToken;
+use tokio_stream::Stream;
 
-use crate::authentication::error::AuthenticationError;
+use crate::{
+    authentication::{AuthenticationLayer, error::AuthenticationError},
+    uds::Uds,
+};
 
-pub(crate) async fn is_authenticated(
-    stream: UnixStream,
-    _shutdown_token: CancellationToken,
-) -> Result<UnixStream, AuthenticationError> {
-    Ok(stream)
+pub(crate) type StreamItem = tokio::net::UnixStream;
+
+// Not implemented yet, implicit is to consider it authenticated
+pub(crate) async fn is_authenticated(_stream: &mut UnixStream) -> Result<(), AuthenticationError> {
+    Ok(())
+}
+
+pub(crate) fn incoming(uds: Uds) -> impl Stream<Item = std::io::Result<StreamItem>> {
+    let auth_layer = AuthenticationLayer::new(uds);
+    auth_layer.stream()
 }