+++ nym-vpn-core/crates/nym-windows/src/process.rs
@@ -5,14 +5,44 @@
 use std::{
     ffi::{CStr, c_char},
     mem,
+    os::windows::io::AsRawHandle,
 };
+use tokio::net::windows::named_pipe::NamedPipeServer;
+use verifysign::CodeSignVerifier;
 use windows::Win32::{
     Foundation::{CloseHandle, ERROR_NO_MORE_FILES, HANDLE},
-    System::Diagnostics::ToolHelp::{
-        CREATE_TOOLHELP_SNAPSHOT_FLAGS, CreateToolhelp32Snapshot, MODULEENTRY32, Module32First,
-        Module32Next, PROCESSENTRY32W, Process32FirstW, Process32NextW,
+    System::{
+        Diagnostics::ToolHelp::{
+            CREATE_TOOLHELP_SNAPSHOT_FLAGS, CreateToolhelp32Snapshot, MODULEENTRY32, Module32First,
+            Module32Next, PROCESSENTRY32W, Process32FirstW, Process32NextW,
+        },
+        Pipes::GetNamedPipeClientProcessId,
     },
 };
+
+/// Errors returned by some functions in this module.
+#[derive(thiserror::Error, Debug)]
+pub enum Error {
+    /// Error returned from `GetNamedPipeClientProcessIdFailed`
+    #[error("failed to get client PID")]
+    GetNamedPipeClientProcessId(#[source] windows::core::Error),
+
+    /// Could not get verification data for the given PID
+    #[error("failed to get the verifier for the client PID: {0:?}")]
+    PidVerifier(verifysign::Error),
+
+    /// Certificate signature verification failure
+    #[error("failed to verify the certificate signatures {0:?}")]
+    Verification(verifysign::Error),
+
+    /// No serial number found in signature
+    #[error("no serial number found in signature")]
+    NoSerialNumber,
+
+    /// Serial number mismatch
+    #[error("serial number of certificate signature is not from Nym")]
+    SerialNumberMismatch,
+}
 
 /// A snapshot of process modules, threads, and heaps
 pub struct ProcessSnapshot {
@@ -157,5 +187,74 @@ impl Iterator for ProcessSnapshotEntries<'_> {
             pid: self.temp_entry.th32ProcessID,
             parent_pid: self.temp_entry.th32ParentProcessID,
         }))
+    }
+}
+
+/// Information about a client process connected to the named pipe server.
+pub struct ClientProcess {
+    /// Process ID
+    pub pid: u32,
+}
+
+impl TryFrom<&NamedPipeServer> for ClientProcess {
+    type Error = Error;
+
+    fn try_from(named_pipe_server: &NamedPipeServer) -> Result<Self, Self::Error> {
+        let mut pid: u32 = 0;
+
+        unsafe { GetNamedPipeClientProcessId(HANDLE(named_pipe_server.as_raw_handle()), &mut pid) }
+            .map_err(Error::GetNamedPipeClientProcessId)?;
+
+        Ok(Self { pid })
+    }
+}
+
+impl ClientProcess {
+    pub fn verify_certificate_signature(
+        &self,
+        nym_certificate_serial_number: String,
+    ) -> Result<(), Error> {
+        let verifier = CodeSignVerifier::for_pid(self.pid).map_err(Error::PidVerifier)?;
+        let context = verifier.verify().map_err(Error::Verification)?;
+        let signer_serial = context.serial().ok_or(Error::NoSerialNumber)?;
+        if signer_serial != nym_certificate_serial_number {
+            Err(Error::SerialNumberMismatch)
+        } else {
+            Ok(())
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::path::PathBuf;
+
+    use tokio::net::windows::named_pipe::{ClientOptions, ServerOptions};
+
+    use super::*;
+
+    #[tokio::test]
+    async fn deny_unsigned() {
+        let pipe_name = PathBuf::from(r"\\.\pipe\nym-vpn-testing");
+        let server = ServerOptions::new()
+            .first_pipe_instance(true)
+            .create(&pipe_name)
+            .unwrap();
+        let _client = ClientOptions::new()
+            .read(true)
+            .write(true)
+            .open(&pipe_name)
+            .unwrap();
+
+        let client_process = ClientProcess::try_from(&server).unwrap();
+        let result = client_process.verify_certificate_signature("dummy".to_string());
+
+        // needed for some reason on windows, otherwise the test blocks
+        drop(server);
+
+        assert!(matches!(
+            result.unwrap_err(),
+            Error::Verification(verifysign::Error::Unsigned)
+        ))
     }
 }